task(SDK-5045) - Moves classes to media package
